improved rendering square logic

+++ exercise/src/draw/draw.h
@@ -8,6 +8,6 @@ void	draw_circle(t_vct ctr, int size, int color);
 
 void	draw_quad(t_quad quad, int color);
 
-void	draw_square(t_vct center, t_vct size, float color);
+void	draw_square(t_vct center, t_vct size, int color);
 
 #endif
